Restructure the clg source files into the clp namespace.
